+++ Common/Integrations/Automate/IAutomateApi.cs
@@ -2,15 +2,11 @@ using System.Collections.Generic;
 using Microsoft.Xna.Framework;
 using StardewValley;
 
-namespace Pathoschild.Stardew.Automate
+namespace Pathoschild.Stardew.Common.Integrations.Automate
 {
-    /// <summary>The API which lets other mods interact with Automate.</summary>
-    public interface IAutomateAPI
+    /// <summary>The API provided by the Automate mod.</summary>
+    public interface IAutomateApi
     {
-        /// <summary>Add an automation factory.</summary>
-        /// <param name="factory">An automation factory which construct machines, containers, and connectors.</param>
-        void AddFactory(IAutomationFactory factory);
-
         /// <summary>Get the status of machines in a tile area. This is a specialised API for Data Layers and similar mods.</summary>
         /// <param name="location">The location for which to display data.</param>
         /// <param name="tileArea">The tile area for which to display data.</param>